\name{print.constraint}
\alias{print.constraint}
\title{print constraint}
\description{print method that shows each variable and constant}
\usage{
\method{print}{constraint}(x, ...)
}
\arguments{
  \item{x}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





