\name{print.variable}
\alias{print.variable}
\title{print variable}
\description{print method that shows variable indices and coefficient}
\usage{
\method{print}{variable}(x, ...)
}
\arguments{
  \item{x}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





