\name{run.quadmod}
\alias{run.quadmod}
\title{run quadmod}
\description{Main function of the package. Should allow definition of the
objective and constraints with a modeling language, then analyze
and eliminate constraints and variables, then run solve.QP.}
\usage{run.quadmod(vars, objective, constraints)}
\arguments{
  \item{vars}{List of variables.}
  \item{objective}{Some description of the objective function.}
  \item{constraints}{List of constraints.}
}



\author{Toby Dylan Hocking}





