\name{sum.lincomblist}
\alias{sum.lincomblist}
\title{sum lincomblist}
\description{collapse a list of linear combinations into a single linear combination}
\usage{
\method{sum}{lincomblist}(lcs, ...)
}
\arguments{
  \item{lcs}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





