\name{>=.lincomblist}
\alias{>=.lincomblist}
\title{>= lincomblist}
\description{make a list of inequality constraints from a list of linear combinations}
\usage{
\method{>=}{lincomblist}(lcs, constant)
}
\arguments{
  \item{lcs}{
}
  \item{constant}{
}
}



\author{Toby Dylan Hocking}





