\name{quadmod-package}
\alias{quadmod-package}
\docType{package}
\title{Quadratic programming modeling language}
\description{An intuitive modeling language for quadratic programs,
and functions to generate the standard form matrices required
to use solve.QP().}
\details{
\tabular{ll}{Package: \tab quadmod\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab Quadratic programming modeling language\cr
Suggests: \tab quadprog\cr
URL: \tab http://quadmod.r-forge.r-project.org\cr}
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}

\keyword{ package }


