\name{quadmod-package}
\alias{quadmod-package}
\alias{quadmod}
\docType{package}
\title{Quadratic programming modeling language}
\description{An intuitive modeling language for quadratic programs,
and functions to generate the standard form matrices required
to use solve.QP().}
\details{
\tabular{ll}{Package: \tab quadmod\cr
Maintainer: \tab Toby Dylan Hocking <toby@sg.cs.titech.ac.jp>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 2013.5.27\cr
License: \tab GPL-3\cr
Title: \tab Quadratic programming modeling language\cr
Depends: \tab quadprog, lpSolveAPI\cr
Suggests: \tab ggplot2\cr
URL: \tab http://quadmod.r-forge.r-project.org\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }


